Turhia renderöintejä pois kuplakaaviosta
